wind up

+++ lib/UI/FacultyInfo/Widgets/barGraph.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 class BarGraph extends StatefulWidget {
   String title;
-  int prof;
+  double prof;
   BarGraph({this.title, this.prof});
   @override
   _BarGraphState createState() => _BarGraphState();
@@ -40,7 +40,7 @@ class _BarGraphState extends State<BarGraph> {
             width: widget.prof* 0.9,
           ),
           Text(
-            '  ${widget.prof} ${widget.title}',
+            '  ${widget.prof.floor()} ${widget.title}',
             style:
                 TextStyle(color: Colors.black54, fontWeight: FontWeight.w600),
           )
